wip: decipher almost works
Need a way to detect repetition. But - I can't really see how that is
possible.. If the message is longer than the secret, there's really no
way of knowing if the secret is 'scones' or 'sconessc'.

For example, if the secret is 'a' or 'aaaaa' is pretty much impossible
to understand?